Thumbnail and embedded file services implemented.
Right, I've got the thumbnails in. Thumbnails work for all document types, not just images. I've also put in some code to create a PDF version of the document, which can be embedded with PDF.JS. This code is nearly complete and I think it's going to be pretty useful - especially on mobile devices which may not be able to read a particular document type (we simply send them the PDF version we have)

New things you need to install:

The latest version of OpenOffice
Apache Maven
brew install ghostscript

In order to generate a PDF version of files, I've had to write a java service. Don't worry, if you're not running it, you just don't get thumbnails of docs, etc but it otherwise works well. To start the conversion service, first run the start-open-office.sh, then start the service by cd'ing to converter-service/converter and running "mvn camel:run"

As I said, it's not the end of the world if you would rather not bother with the java bits for now