Sanitize timezone just in case.
ElasticSearch seems to validate it and throw an error because I get a 400 error
if I pass in gibberish. The chats lookup that goes via mongo (via the URL in
the address bar) appears to deal with it by just going with utc or something
and the only way you can get to a broken elastic search one is by some console
trickery, but just in case something parses to a well crafted object I'm
forcing it to a string according to Andy's recommendation.