Fix homepage mixpanel on dev and beta.
window.mixpanel always exists, even when you haven't initialised it. If you
don't, then it isn't enough to see if window.mixpanel is truthy before trying
to call a method. Rather check for the method you're trying to call. This is
because those methods only exist after initialisation.

We don't pass a "library_name" to any of the calls to mixpanel.init, so we
can't check for mixpanel.library_name either. Therefore this just seems like
the safest fix  to me.