Archive index in sitemap, split into index.
This puts 50k rooms per sitemap file and then lists all the sitemap files into
a sitemap index. So sitemap.xml actually becomes the sitemap index file.

Next step is to link to all the days from the index page. It could even be done
in a <noscript> block if we don't want to turn this into a design exercise.