Add archive links to the archive index page.
The only problem here is that I don't know the timezone yet because elsewhere
we get it using javascript and then pop it in a URL. So if you go to that URL,
then the heatmap and archive links will be out of sync.

An obvious solution would be to just make a separate archive page for search
engines that always uses UTC and then just link to that (rather than the
heatmap one) from the sitemap.

This still needs cache headers. I think all days before today (or more than 12
or 24 hours out to account for timezone weirdness?) can just cache forever and
the current day much shorter. The index page (heatmap one or the proposed new
one) can cache for a day as a new link will only appear once a day max.

But this is the basic idea, anyway.