Fix room activity badges not updating
If we consume our data within the primary/favourite collections from a
static snapshot we dis-associate our models with the models from the room
collection. This results in updates not being passed up into the menu.
To combat this I have replaced the filters in the room-menu-model which
means the models passed to the primary/favourite collections are the ones
from the room collection and as such receive run-time updates via
Halley.