Avoid redirect loop even when forced token
If you are using the desktop app, we force feed the token given via the `Authorization: Bearer xxx` header.
But if that token is invalid or revoked, it will result in a infinite redirect loop (ends in a browser too_many_redirects).

With these changes, we will now redirect you back to `/login` `/login/token-revoked` appropriately and
not try to redirect again even if you have a invalid token present.