Suggest emojis only when two chars have been typed
This mirrors the behavior of Slack, and makes it easier to write
text with lines that end with an emoticon. Previously, ending a line (or
submitting a comment) with :p by pressing Enter would change the :p in
to :package:, which is not what the user intended.

Note that with this change, a single : will no longer produce an emoji
auto-complete. This behavior could be changed by adding a ? behind the
capture group to indicate that no text following the : is also
acceptable.

Fixes #1939.