Resolve "Permalinks in thread message feed"
There is one refactoring commit that splits the `findChatMessagesInTroupe` method into two independent methods 7b2a48fb56d26d4f2ba78f1cd6b790d1a9b627b5. It's going to be best to look at this one in isolation because it is not directly related to the permalink feature.

The rest of the MR is implementing addition to permalink functionality: when permalink points to the child message, Gitter will highlight parent message, opens TMF and highlights the child message.

## TODO:

* [x] Clicking the timestamp in TMF populates location bar with permalink
* [x] Server uses parent message when `aroundId` belongs to child message and adds the child message to the response
* [x] Client checks whether the permalinked message is a child message and if yes, it highlights parent, opens TMF, and highlights child
* [x] TMF scrolls to the highlighted child