Remove extra click because it can also fail the test if unable to click twice
The extra click was added because the test would sometimes just fail to open the unread -> thread area.
But now I have also seen it the other way where it opens properly on the first click and fails
on the second click because the banner is now gone. So I am opting for the proper expected first click route.