Remove community creation invite step
Fix https://gitlab.com/gitlab-org/gitter/webapp/issues/2103

Invites can be done after the fact from the room. Nothing special about adding people there

Sometimes the invites can fail and the community creates sucessfully but
one of the invite fails so the UI shows an error which just confuses people.

We could make the invite errors more clear in the creation process but it's
simpler just to remove the invite step.

It also removes a barrier to entry to just create a community (less to think about).