Fix GitLab group and user service to return individualized responses
Previously, we would serve whatever response was first in the cache

We were using `options.getInstanceId` to cache based on the user token
but the cache wrapper didn't use it for individual functions.

This only affected the GitLab group and user service because we only cache some functions.
For the GitLab issuable and state service, we cache the whole class which already uses
`options.getInstanceId` as expected.

I discovered this while working on https://gitlab.com/gitlab-org/gitter/webapp/-/merge_requests/1757
and testing with 2 GitLab accounts, I noticed I got the same response from the orgs
endpoint when creating a community