Fix UI bug by properly describing app behavior
As noted here:
https://gitlab.com/gitlab-org/gitter/webapp/issues/2456
https://gitlab.com/gitlab-org/gitter/webapp/issues/1205

"Unread item counts" is tied to the setting for 'mentions and announcements', which means
that all unread messages will eventually trigger emails to be sent.  This fixes that bug
by conveying to the user that they are going to get emails in both cases: either
immediately, or in batches.