Update gittertestbot email to match in tests
Recently updated the email for gittertestbot so anyone with Gitter address can access.

For some reason, it wasn't finding any commits when filtering by commit author `gittertestbot`.
I needed to add a new commit to one of the repos so it would find a commit and so that the email would match
with the auto discovery.